B-02397 Admin 2.0: Accounts: New User: Consolidate user store step and profile step
We want to remove the user store step. Instead we'll add the user store list to the profile step.

When the user selects a user store the profile form should be shown in the same step.

Once a user store is selected it should not be possible to select user store.