B-02398 Admin 2.0: Display name must be autogenerated when typing
In the profile step of the user wizard, updating the name fields should trigger an update of the display-name field on top.
The display name should be autogenerated from the following fields: prefix, firstName, middleName, lastName, suffix

The display name should be updated live while the user types in any of the "name" fields.