B-02517 Accounts: New Icon for Roles (Built-in Group types)
We need to add a different Icon for roles (built-in groups) from the one used for standard groups.

Use the icon "businessmen"

Remember to replace in the following locations:

- Account lists
- Button (group selector, small details)
- Medium button (medium details)
- Visitation card (dialogues)
- Group Details big
- Group edit/new form

We will use the name "ROLE" for these group types in the future, so that may be a smart name to use in the code for this!