B-02546 Role forms
For the moment we will treat all roles very similar to "groups", we will get back to special treatmen of things like anonymous user later on!

We will need to do some special handling when it comes to editing roles

The following roles shall not be possible to edit (i.e. remove edit button)

* Anonymous (as user)
* Authenticated Users (as Group)

NB: If there is a separate role for Enterprise Admin (user) in comparison to "Enterprise Administrators" (group) the user role shall not be possible to edit

For now, we will not do anything special with the details view (separate task)

For the other roles we will do as following with the form

The role form should be the same form as for groups (or based upon it)?

Step 1:

Remove the Entire step 1 (conditional?) - none of these fields are possible to alter for roles anyway

Step 2: (Now step 1)

Add a short description text at the top of the form (different text per role type)

Finally, ensure that editable roles are possible to save!