B-02554 Tab opening responsiveness
When opening new tabs (edit or view) there is a lag (especially visible with bad connections) - and it appears that the tab is not opened until it is fully loaded?

We want to improve responsiveness by opening the tab instantly (with title and icons) and then display a loading splash inside the tab until the content has loaded.